remove unused handlers
