--- x/monitoringc/keeper/monitoring.go
@@ -1,72 +1,0 @@
-package keeper
-
-import (
-	"errors"
-	spntypes "github.com/tendermint/spn/pkg/types"
-
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	channeltypes "github.com/cosmos/ibc-go/v2/modules/core/04-channel/types"
-	"github.com/tendermint/spn/x/monitoringc/types"
-)
-
-
-// OnRecvMonitoringPacket processes packet reception
-func (k Keeper) OnRecvMonitoringPacket(
-	ctx sdk.Context,
-	packet channeltypes.Packet,
-	data spntypes.MonitoringPacket,
-	) (packetAck types.MonitoringPacketAck, err error) {
-	// validate packet data upon receiving
-	if err := data.ValidateBasic(); err != nil {
-		return packetAck, err
-	}
-
-	// TODO: packet reception logic
-
-	return packetAck, nil
-}
-
-// OnAcknowledgementMonitoringPacket responds to the the success or failure of a packet
-// acknowledgement written on the receiving chain.
-func (k Keeper) OnAcknowledgementMonitoringPacket(
-	ctx sdk.Context,
-	packet channeltypes.Packet,
-	data spntypes.MonitoringPacket,
-	ack channeltypes.Acknowledgement,
-	) error {
-	switch dispatchedAck := ack.Response.(type) {
-	case *channeltypes.Acknowledgement_Error:
-
-		// TODO: failed acknowledgement logic
-		_ = dispatchedAck.Error
-
-		return nil
-	case *channeltypes.Acknowledgement_Result:
-		// Decode the packet acknowledgment
-		var packetAck types.MonitoringPacketAck
-
-		if err := types.ModuleCdc.UnmarshalJSON(dispatchedAck.Result, &packetAck); err != nil {
-			// The counter-party module doesn't implement the correct acknowledgment format
-			return errors.New("cannot unmarshal acknowledgment")
-		}
-
-		// TODO: successful acknowledgement logic
-
-		return nil
-	default:
-		// The counter-party module doesn't implement the correct acknowledgment format
-		return errors.New("invalid acknowledgment format")
-	}
-}
-
-// OnTimeoutMonitoringPacket responds to the case where a packet has not been transmitted because of a timeout
-func (k Keeper) OnTimeoutMonitoringPacket(
-	ctx sdk.Context,
-	packet channeltypes.Packet,
-	data spntypes.MonitoringPacket,
-	) error {
-
-	// TODO: packet timeout logic
-
-	return nil
-}
